# Rename IFieldResolver.Resolve to ResolveAsync and return Task<object>