--- a/src/GraphQL.StarWars/StarWarsData.cs
+++ b/src/GraphQL.StarWars/StarWarsData.cs
@@ -1,4 +1,4 @@
-using System.Collections.Generic;
+ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using GraphQL.StarWars.Types;