--- a/src/GraphQL.Tests/Builders/FieldBuilderTests.cs
+++ b/src/GraphQL.Tests/Builders/FieldBuilderTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using System.Threading.Tasks;
 using GraphQL.StarWars.Types;
 using GraphQL.Types;
 using Shouldly;
@@ -53,7 +54,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void should_return_the_right_type()
+        public async Task should_return_the_right_type()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -67,8 +68,8 @@ namespace GraphQL.Tests.Builders
             fields[0].Type.ShouldBe(typeof(StringGraphType));
 
             var context = new ResolveFieldContext();
-            fields[0].Resolver.Resolve(context).GetType().ShouldBe(typeof(string));
-            fields[0].Resolver.Resolve(context).ShouldBe("SomeString");
+            (await fields[0].Resolver.ResolveAsync(context)).GetType().ShouldBe(typeof(string));
+            (await fields[0].Resolver.ResolveAsync(context)).ShouldBe("SomeString");
         }
 
         [Fact]
@@ -140,7 +141,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void can_determine_whether_argument_exists_in_resolver()
+        public async Task can_determine_whether_argument_exists_in_resolver()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<IntGraphType>()
@@ -160,7 +161,7 @@ namespace GraphQL.Tests.Builders
             field.Arguments[0].Description.ShouldBe("desc1");
             field.Arguments[0].Type.ShouldBe(typeof(StringGraphType));
             field.Arguments[0].DefaultValue.ShouldBe("12345");
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>
                 {
@@ -170,7 +171,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void getting_unspecified_argument_in_resolver_yields_null()
+        public async Task getting_unspecified_argument_in_resolver_yields_null()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -182,7 +183,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>(),
                 FieldDefinition = field
@@ -190,7 +191,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void getting_unspecified_argument_in_resolver_yields_overridden_default_value()
+        public async Task getting_unspecified_argument_in_resolver_yields_overridden_default_value()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -202,7 +203,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>(),
                 FieldDefinition = field
@@ -210,7 +211,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void can_get_nullable_argument_with_value()
+        public async Task can_get_nullable_argument_with_value()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -222,7 +223,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>
                 {
@@ -233,7 +234,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void can_get_nullable_argument_with_null_value()
+        public async Task can_get_nullable_argument_with_null_value()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -245,7 +246,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>
                 {
@@ -256,7 +257,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void can_get_nullable_argument_missing_value()
+        public async Task can_get_nullable_argument_missing_value()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -268,7 +269,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>(),
                 FieldDefinition = field
@@ -276,7 +277,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void can_get_enum_argument()
+        public async Task can_get_enum_argument()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -288,7 +289,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>
                 {
@@ -299,7 +300,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void can_get_enum_argument_with_overriden_default_value()
+        public async Task can_get_enum_argument_with_overriden_default_value()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -311,7 +312,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>(),
                 FieldDefinition = field
@@ -319,7 +320,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void can_get_list_argument()
+        public async Task can_get_list_argument()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -331,7 +332,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>
                 {
@@ -342,7 +343,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void can_get_collection_argument()
+        public async Task can_get_collection_argument()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -354,7 +355,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>
                 {
@@ -366,7 +367,7 @@ namespace GraphQL.Tests.Builders
 
 
         [Fact]
-        public void getting_specified_argument_in_resolver_overrides_default_value()
+        public async Task getting_specified_argument_in_resolver_overrides_default_value()
         {
             var objectType = new ObjectGraphType();
             objectType.Field<StringGraphType>()
@@ -378,7 +379,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Arguments = new Dictionary<string, object>
                 {
@@ -389,7 +390,7 @@ namespace GraphQL.Tests.Builders
         }
 
         [Fact]
-        public void can_access_object()
+        public async Task can_access_object()
         {
             var objectType = new ObjectGraphType<int>();
 
@@ -401,7 +402,7 @@ namespace GraphQL.Tests.Builders
                 });
 
             var field = objectType.Fields.First();
-            field.Resolver.Resolve(new ResolveFieldContext
+            await field.Resolver.ResolveAsync(new ResolveFieldContext
             {
                 Source = 12345
             });