… src/GraphQL.ApiTests/ApiApprovalTests.PublicApi.approved.txt
… src/GraphQL/Utilities/ISchemaNodeVisitor.cs
… src/GraphQL/Utilities/SyncToAsyncResolverAdapter.cs