Based on the comment by @joemcbride that IValidationRule.Validate should become ValidateAsync and return a `Task<INodeVisitor>`, I suggest that IFieldResolver.Resolve should become ResolveAsync and return a `Task<object>`.  `IFieldResolver<T>.Resolve` returns a `Task<T>`.  This pull request represents the necessary changes.

Of note, field builders already have Resolve and ResolveAsync methods, so they shouldn't change.  However, many of the asynchronous samples in the tests and examples would build the asynchronous field with Field() rather than FieldAsync().  Using FieldAsync is now required for asynchronous fields.  This would be the biggest breaking change.

This cleans up some of the mess in ExecutionStrategy.ExecuteNodeAsync, which was relying on returned type inspection to determine if it needed to await on the return value or not, and then having to make a call to `((dynamic)task).Result` (within TaskExtensions.GetResult) to retrieve the value.

The internal SyncToAsyncResolverAdapter has been removed, as it serves no purpose.

The DelegateFieldModelBinderResolver still inspects the return type and awaits on it if it is a Task.  Personally, I would remove this class from the base GraphQL library, as it promotes non-type-safe coding, and any third party can reproduce this functionality extremely easily.  Perhaps if I knew how it was being used, I would have a different opinion, but in any case, it's still there.

Finally, the AccessorFieldResolver also still inspects the return type and awaits on it if it is a Task.  This seemed to make sense given the usage.

References: #1321 and #1140