@Shane32 2019-10-31
One other note: these changes should have no effect on anyone using synchronous field resolvers using the basic Field() method or FieldBuilder.Resolve() methods, or on anyone using asynchronous field resolvers using the basic FieldAsync() method or FieldBuilder.ResolveAsync() methods.  It's only when (a) defining a custom IFieldResolver, or (b) returning an asynchronous result using the Field() or FieldBuilder.Resolve() methods.