@sungam3r 2019-10-31
I looked at the changes, not in great detail, but the overall picture is clear. I can’t say that the code has become simpler. Also, I can’t say that it brings only positive changes. With these changes, you force all consumers (explicitly or implicitly) to use Task to wrap values. That is, **each field** will now be forced to be wrapped in the Task when the value is received from resolver. This is a definite hit on performance.

> Based on the comment by @joemcbride that IValidationRule.Validate should become ValidateAsync and return a Task, I suggest that IFieldResolver.Resolve should become ResolveAsync and return a Task. 

There is a big difference between the two examples. Field resolver is not required to return Task with some object. It can return both it and just some object (synchronous return). `ExecutionStrategy` checks it and do approprite await if needed:
```c#
 if (result is Task task)
                {
                    await task.ConfigureAwait(false);
                    result = task.GetResult();
                }
```

However, I agree that in general, you can get lost in the variety of possible signatures for determining the resolver for the field.

Relates to #1146 

Also I remember that I saw another issue about it would be nice to refactor 6 different ways of setting field, but I can not find it.