@sungam3r 2019-10-31
> One other note: these changes should have no effect on anyone using synchronous field resolvers using the basic Field() method or FieldBuilder.Resolve() methods

I don't understand. How will this not affect them if another type of return is now required?