@Shane32 2019-10-31
> I don't understand. How will this not affect them if another type of return is now required?

To recap, using Field() or FieldBuilder.Resolve() accepts a synchronous Func<> and wraps the resolve method parameter in a FuncFieldResolver, which handles wrapping the result into a Task<> via Task.FromResult.  (Implicit use of Task.FromResult.)  When using FieldAsync(), the resolve parameter is wrapped in an AsyncFieldResolver which passes the Task through.

So I guess I didn't mean to say that these changes wouldn't affect performance for synchronous fields, but rather that it would not typically be a breaking change in the way the fields are defined, nor in the ultimate execution of the query.

> This is a definite hit on performance.

I agree, there is a hit on performance.  However, .Net Core is optimized for this behavior.  Based on this website: https://github.com/dotnet/corefx/issues/30655 , when using ConfigureAwait(false), it is about 10 nanoseconds.  My own testing showed it to be about 16 nanoseconds, as opposed to 5 nanoseconds when using the prior code design (the 'is' operator is quite slow).  This means that for 10,000 executions of the resolve function, it's about 0.11ms slower.  Not really enough to quibble about.

Personally, I don't think it's a big deal either way.  If I was picking, although I normally go for every performance enhancement possible, in this case I would probably choose tightly bound interfaces over using 'is' and dynamically invoking members, as the performance impact is negligible.  But I only posted it because of the goals @joemcbride had for 3.0 in relation to IValidationRule, and felt this would be a change in the same vein.  If the change is desired, it should probably make it into 3.0, as it is a breaking change and would not make sense to introduce at a later time.