@Shane32 2020-01-13
I would on principle, and I also thought it would match the change in the IValidationRule.  But I really don't think it matters too much, and we can just close it if you think it's better to leave as-is.