@sungam3r 2020-01-13
My intuition says it's worth leaving everything as it is.