@Shane32 2020-02-07
As all synchronous field resolvers are required to use Task.FromResult to return the value, there are additional objects created and disposed of, consisting of a slight performance decrease, as shown here:

```BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18362
Intel Core i7-7740X CPU 4.30GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT

*** BEFORE ***

|        Method |        Mean |    Error |   StdDev |    Gen 0 |   Gen 1 | Gen 2 |  Allocated |
|-------------- |------------:|---------:|---------:|---------:|--------:|------:|-----------:|
| Introspection | 1,516.36 us | 9.253 us | 8.202 us | 201.1719 | 83.9844 |     - | 1050.88 KB |
|          Hero |    31.56 us | 0.266 us | 0.249 us |   6.2256 |       - |     - |   25.55 KB |

*** AFTER ***

|        Method |        Mean |     Error |    StdDev |    Gen 0 |    Gen 1 | Gen 2 |  Allocated |
|-------------- |------------:|----------:|----------:|---------:|---------:|------:|-----------:|
| Introspection | 1,635.75 us | 13.142 us | 12.293 us | 222.6563 | 109.3750 |     - | 1134.81 KB |
|          Hero |    32.10 us |  0.266 us |  0.249 us |   6.2866 |   0.0610 |     - |   25.77 KB |```