@sungam3r 2020-02-08
8% memory growth is quite noticeable and I think it will be all the more noticeable the more leaf nodes in the schema for which the results are wrapped into a `Task`.