@sungam3r 2020-02-08
Also #1146 should show the opposite effect - memory desrease. Then we can compare two extremes - when everything wrapped into a `Task` and when nothing wrapped.